v2

+++ src/app/model/observable-array.ts
@@ -1,57 +1,93 @@
-import {BehaviorSubject} from "rxjs/BehaviorSubject";
-import {Observable} from "rxjs/Observable";
-import {ArrayService} from "../services/utils/array.service";
+import {BehaviorSubject} from 'rxjs/BehaviorSubject';
+import {Observable} from 'rxjs/Observable';
+import {ArrayService} from '../services/utils/array.service';
 
 
 export class ObservableArray<T> {
-    public readonly items$: Observable<T[]>;
-    public readonly itemsSource: BehaviorSubject<T[]>;
+    protected readonly itemsSource: BehaviorSubject<T[]>;
+    protected readonly beforeNextSource: BehaviorSubject<T[]>;
 
 
     constructor(itemList: T[]) {
-        this.itemsSource = new BehaviorSubject<T[]>(itemList)
-        this.items$ = this.itemsSource.asObservable();
+        this.itemsSource = new BehaviorSubject<T[]>(itemList);
+        this.beforeNextSource = new BehaviorSubject<T[]>(undefined);
+    }
+
+
+    get beforeNext$(): Observable<T[]> {
+        return this.beforeNextSource.asObservable();
+    }
+
+
+    get items$(): Observable<T[]> {
+        return this.itemsSource.asObservable();
+    }
+
+
+    get lastItem$(): Observable<T> {
+        return this.items$
+            .map(items => (items && items.length > 0) ? items[items.length - 1] : undefined);
     }
 
 
     public replaceList(itemList: T[]) {
+        this.beforeNextSource.next(itemList);
         this.itemsSource.next(itemList);
     }
 
 
     public clear() {
-        this.itemsSource.next([]);
+        const itemList = [];
+        this.beforeNextSource.next(itemList);
+        this.itemsSource.next(itemList);
     }
 
 
     public push(item: T) {
         const itemList = this.itemsSource.getValue();
-        itemList.push(item);
-        this.itemsSource.next(itemList);
+
+        if (itemList) {
+            itemList.push(item);
+            this.beforeNextSource.next(itemList);
+            this.itemsSource.next(itemList);
+        }
     }
 
 
     public insert(item: T, index: number) {
         const itemList = this.itemsSource.getValue();
-        ArrayService.insertAt(itemList, index, item);
-        this.itemsSource.next(itemList);
+
+        if (itemList && index >= 0) {
+            ArrayService.insertAt(itemList, index, item);
+            this.beforeNextSource.next(itemList);
+            this.itemsSource.next(itemList);
+        }
     }
 
 
     public remove(item: T) {
         const itemList = this.itemsSource.getValue();
-        ArrayService.removeFromArray(itemList, item);
-        this.itemsSource.next(itemList);
+
+        if (itemList) {
+            ArrayService.removeFromArray(itemList, item);
+            this.beforeNextSource.next(itemList);
+            this.itemsSource.next(itemList);
+        }
     }
 
 
     public replace(index: number, item: T) {
         const itemList = this.itemsSource.getValue();
-        itemList[index] = item;
-        this.itemsSource.next(itemList);
+
+        if (itemList && index >= 0 && index < itemList.length) {
+            itemList[index] = item;
+            this.beforeNextSource.next(itemList);
+            this.itemsSource.next(itemList);
+        }
     }
 
 
+    // TODO
     public indexOf(item: T): number {
         const itemList = this.itemsSource.getValue();
         return itemList.indexOf(item);
